internal/contour: split metrics calculation into its own file
