+++ internal/contour/metrics.go
@@ -1,0 +1,54 @@
+// Copyright © 2018 Heptio
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package contour contains the translation business logic that listens
+// to Kubernetes ResourceEventHandler events and translates those into
+// additions/deletions in caches connected to the Envoy xDS gRPC API server.
+package contour
+
+import (
+	"github.com/heptio/contour/internal/dag"
+	"github.com/heptio/contour/internal/metrics"
+)
+
+func calculateIngressRouteMetric(statuses map[dag.Meta]dag.Status) metrics.IngressRouteMetric {
+	metricTotal := make(map[metrics.Meta]int)
+	metricValid := make(map[metrics.Meta]int)
+	metricInvalid := make(map[metrics.Meta]int)
+	metricOrphaned := make(map[metrics.Meta]int)
+	metricRoots := make(map[metrics.Meta]int)
+
+	for _, v := range statuses {
+		switch v.Status {
+		case dag.StatusValid:
+			metricValid[metrics.Meta{VHost: v.Vhost, Namespace: v.Object.GetNamespace()}]++
+		case dag.StatusInvalid:
+			metricInvalid[metrics.Meta{VHost: v.Vhost, Namespace: v.Object.GetNamespace()}]++
+		case dag.StatusOrphaned:
+			metricOrphaned[metrics.Meta{Namespace: v.Object.GetNamespace()}]++
+		}
+		metricTotal[metrics.Meta{Namespace: v.Object.GetNamespace()}]++
+
+		if v.Object.Spec.VirtualHost != nil {
+			metricRoots[metrics.Meta{Namespace: v.Object.GetNamespace()}]++
+		}
+	}
+
+	return metrics.IngressRouteMetric{
+		Invalid:  metricInvalid,
+		Valid:    metricValid,
+		Orphaned: metricOrphaned,
+		Total:    metricTotal,
+		Root:     metricRoots,
+	}
+}
